make JS accept floating-point pause values.
git-svn-id: http://svn.warlock.cc/svn/warlock/warlock2/trunk@1455 bb130017-240f-0410-9613-9c2c827820fb

+++ core/cc.warlock.core.script/src/main/cc/warlock/core/script/javascript/JavascriptCommands.java
@@ -101,7 +101,7 @@ public class JavascriptCommands {
 		commands.move(direction);
 	}
 
-	public void pause(int seconds) {
+	public void pause(double seconds) {
 		checkStop();
 		
 		commands.pause(seconds);
